+++ src/nextsync/__init__.py
@@ -2,4 +2,4 @@
 
 APP_ID = "io.github.gnacho.nextsync"
 APP_NAME = "NextSync"
-VERSION = "0.3.0"
+VERSION = "0.4.0"